requirements.txt
python-telegram-bot==21.10
openai>=1.68.2
python-dotenv==1.0.1
pymupdf==1.25.3
python-docx==1.1.2
pandas==2.2.3
openpyxl==3.1.5
numpy==2.2.3
httpx==0.28.1
